Renamed a TOC definition name
