+++ src/stories/Dummy.stories.tsx
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
 import Dummy from '~/Dummy';
-import { IInteractions } from './toc';
+import { IInteractions } from './stories.toc';
 
 /** Interactions definition for BackstopJS. */
 export const interactions: IInteractions = [];
